Refactor: Extract loadSubmoduleFixtures to shared utility class
