+++ src/test/groovy/me/biocomp/hubitat_ci/ExistingAppScriptsTest.groovy
@@ -8,6 +8,7 @@ import me.biocomp.hubitat_ci.api.common_api.Location
 import me.biocomp.hubitat_ci.api.common_api.Log
 import me.biocomp.hubitat_ci.app.AppValidator
 import me.biocomp.hubitat_ci.app.HubitatAppSandbox
+import me.biocomp.hubitat_ci.util.SubmoduleFixtureLoader
 import me.biocomp.hubitat_ci.validation.Flags
 import spock.lang.Shared
 import spock.lang.Specification
@@ -140,17 +141,7 @@ class IComfortAppScriptTest extends
         Specification
 {
     @Shared
-    static final List<File> REQUIRED_SUBMODULE_FILES = loadSubmoduleFixtures()
-
-    static List<File> loadSubmoduleFixtures() {
-        def configFile = new File(".submodule-fixtures.txt")
-        if (!configFile.exists()) {
-            throw new FileNotFoundException("Configuration file .submodule-fixtures.txt not found")
-        }
-        return configFile.readLines()
-            .findAll { line -> line && !line.trim().startsWith('#') }
-            .collect { new File(it.trim()) }
-    }
+    static final List<File> REQUIRED_SUBMODULE_FILES = SubmoduleFixtureLoader.loadSubmoduleFixtures()
 
     def setupSpec() {
         def missing = REQUIRED_SUBMODULE_FILES.findAll { !it.exists() }
